refactor(tauri-api): #373 Phase 5 tauri-api.ts を 11 領域別 sub-module に分割
Issue #373 / Phase 5 (横断クリーンアップ): `lib/tauri-api.ts` (328 行) を
領域別に 11 sub-module へ分割し、本体は thin facade に。
move only / 機能変更なし / 公開 API (`api` / `Api` / `isTauri`
/ `RoleProfileSummary`) は 1 文字も変えない。

# 構成

```
lib/tauri-api.ts                 328 → 70 行 (thin facade: import 11 領域 + ping
                                              + isTauri + 自動 bootstrap)
lib/tauri-api/
  app.ts             新規  116 行  app.* IPC namespace + RoleProfileSummary export
                                  + 内部 helper interface 6 種 (TeamMcpMember /
                                  SetupTeamMcpResult / 他)
  git.ts             新規   14 行  git.* (status / diff)
  files.ts           新規   41 行  files.* (list / read / write)
  sessions.ts        新規   10 行  sessions.* (list)
  team-history.ts    新規   20 行  teamHistory.* (list / save / saveBatch / delete)
  handoffs.ts        新規   30 行  handoffs.* (create / list / read / updateStatus)
  dialog.ts          新規   11 行  dialog.* (openFolder / openFile / isFolderEmpty)
  settings.ts        新規   12 行  settings.* (load / save)
  role-profiles.ts   新規    9 行  roleProfiles.* (load / save)
  logs.ts            新規   14 行  logs.* (readTail / openDir)
  terminal.ts        新規   46 行  terminal.* (create / write / resize / kill /
                                  savePastedImage + onData/Exit/SessionId 6 種
                                  + 内部 SavePastedImageResult)
```

# 公開 API の維持

- `import { api } from '../lib/tauri-api'` (FileTreePanel.tsx 等) — 不変
- `import { isTauri } from '../lib/tauri-api'` (ImagePreview.tsx) — 不変
- `import { RoleProfileSummary } from '../lib/tauri-api'` (role-profiles-context.tsx) —
  `pub use ./tauri-api/app` で再 export
- IPC コマンド名 / event 名 / 関数 signature はすべて bit identical
- 自動 bootstrap (window.api への shim 注入) も逐字保持

# 検証結果

- `npm run typecheck`: ✅ green
- 削除 283 行 + 追加 32 行 = 圧縮 -251 行 (本体は 328 → 70 行)、合計は 11 sub-module
  + facade で +28 行 (helper interface の領域分散による軽微な重複)

# Agent Team で並列分担

11 sub-module の draft を 3 人の teammate (app-group / terminal-group
/ data-group extractor) に並列で書かせ、team-lead 側で thin facade を Write して
typecheck 通しを実施。各 sub-module は bit identical で関数本体・doc コメント・
型注釈すべて逐字保持。

Refs #373
