+++ src/renderer/src/lib/tauri-api/app.ts
@@ -1,0 +1,116 @@
+// tauri-api/app.ts — app.* IPC namespace (Phase 5 / Issue #373)
+
+import { invoke } from '@tauri-apps/api/core';
+import type {
+  AppUserInfo,
+  ClaudeCheckResult,
+  SetWindowEffectsResult,
+  ThemeName
+} from '../../../../types/shared';
+
+/** Tauri 側 TeamHub に同期する role profile の要約形 */
+export interface RoleProfileSummary {
+  id: string;
+  labelEn: string;
+  labelJa?: string;
+  descriptionEn: string;
+  descriptionJa?: string;
+  canRecruit: boolean;
+  canDismiss: boolean;
+  canAssignTasks: boolean;
+  /** Leader が team_create_role / team_recruit(role_definition=...) で動的ロールを作れるか */
+  canCreateRoleProfile: boolean;
+  defaultEngine: string;
+  singleton: boolean;
+}
+
+interface TeamMcpMember {
+  agentId: string;
+  role: string;
+  agent: string;
+}
+interface SetupTeamMcpResult {
+  ok: boolean;
+  error?: string;
+  socket?: string;
+  changed?: boolean;
+}
+interface CleanupTeamMcpResult {
+  ok: boolean;
+  error?: string;
+  removed?: boolean;
+}
+interface ActiveLeaderResult {
+  ok: boolean;
+  error?: string;
+}
+interface OpenExternalResult {
+  ok: boolean;
+  error?: string;
+}
+interface TeamHubInfo {
+  socket: string;
+  token: string;
+  bridgePath: string;
+}
+
+export const app = {
+  getProjectRoot: (): Promise<string> => invoke('app_get_project_root'),
+  /** Issue #29: renderer 側で project root が切り替わったとき Rust 側 state を同期する */
+  setProjectRoot: (projectRoot: string): Promise<void> =>
+    invoke('app_set_project_root', { projectRoot }),
+  restart: (): Promise<void> => invoke('app_restart'),
+  setWindowTitle: (title: string): Promise<void> => invoke('app_set_window_title', { title }),
+  checkClaude: (command: string): Promise<ClaudeCheckResult> =>
+    invoke('app_check_claude', { command }),
+  setZoomLevel: (level: number): Promise<void> => invoke('app_set_zoom_level', { level }),
+  /**
+   * Issue #260 PR-1: テーマに応じて OS ネイティブの window effect (Windows: Acrylic /
+   * macOS: vibrancy) を切り替える。Linux 等は no-op (applied=false で返る)。
+   * 引数を `ThemeName` に絞ることで誤った文字列での呼び出しをコンパイル時に弾く。
+   */
+  setWindowEffects: (theme: ThemeName): Promise<SetWindowEffectsResult> =>
+    invoke('app_set_window_effects', { theme }),
+  setupTeamMcp: (
+    projectRoot: string,
+    teamId: string,
+    teamName: string,
+    members: TeamMcpMember[]
+  ): Promise<SetupTeamMcpResult> =>
+    invoke('app_setup_team_mcp', { projectRoot, teamId, teamName, members }),
+  cleanupTeamMcp: (projectRoot: string, teamId: string): Promise<CleanupTeamMcpResult> =>
+    invoke('app_cleanup_team_mcp', { projectRoot, teamId }),
+  setActiveLeader: (teamId: string, agentId?: string | null): Promise<ActiveLeaderResult> =>
+    invoke('app_set_active_leader', { teamId, agentId }),
+  getTeamFilePath: (teamId: string): Promise<string> =>
+    invoke('app_get_team_file_path', { teamId }),
+  getMcpServerPath: (): Promise<string> => invoke('app_get_mcp_server_path'),
+  getTeamHubInfo: (): Promise<TeamHubInfo> => invoke('app_get_team_hub_info'),
+  /** RoleProfile summary を Hub へ同期 (team_list_role_profiles / permissions 検証用) */
+  setRoleProfileSummary: (summary: RoleProfileSummary[]): Promise<void> =>
+    invoke('app_set_role_profile_summary', { summary }),
+  /** recruit を手動キャンセル (timeout 待ち中にユーザーがカードを × で閉じた等) */
+  cancelRecruit: (agentId: string): Promise<void> =>
+    invoke('app_cancel_recruit', { agentId }),
+  /**
+   * `<projectRoot>/.claude/skills/vibe-team/SKILL.md` を書き出す。
+   * setupTeamMcp でも best-effort で実行されるが、Onboarding / 設定 UI から手動で
+   * 強制再配置 (forceOverwrite=true) したい場合のために露出する。
+   */
+  installVibeTeamSkill: (
+    projectRoot: string,
+    forceOverwrite?: boolean
+  ): Promise<{
+    ok: boolean;
+    path?: string;
+    skipped?: boolean;
+    overwritten?: boolean;
+    error?: string;
+  }> =>
+    invoke('app_install_vibe_team_skill', { projectRoot, forceOverwrite: !!forceOverwrite }),
+  getUserInfo: (): Promise<AppUserInfo> => invoke('app_get_user_info'),
+  openExternal: (url: string): Promise<OpenExternalResult> => invoke('app_open_external', { url }),
+  /** Issue #251: OS のファイルマネージャで親フォルダを開き該当ファイルをハイライト */
+  revealInFileManager: (path: string): Promise<OpenExternalResult> =>
+    invoke('app_reveal_in_file_manager', { path })
+};
