+++ src/renderer/src/lib/tauri-api/team-history.ts
@@ -1,0 +1,20 @@
+// tauri-api/team-history.ts — teamHistory.* IPC namespace (Phase 5 / Issue #373)
+
+import { invoke } from '@tauri-apps/api/core';
+import type { TeamHistoryEntry } from '../../../../types/shared';
+
+interface MutationResult {
+  ok: boolean;
+  error?: string;
+}
+
+export const teamHistory = {
+  list: (projectRoot: string): Promise<TeamHistoryEntry[]> =>
+    invoke('team_history_list', { projectRoot }),
+  save: (entry: TeamHistoryEntry): Promise<MutationResult> =>
+    invoke('team_history_save', { entry }),
+  /** Issue #132: 複数チームを 1 IPC + 1 disk write でまとめて保存する */
+  saveBatch: (entries: TeamHistoryEntry[]): Promise<MutationResult> =>
+    invoke('team_history_save_batch', { entries }),
+  delete: (id: string): Promise<MutationResult> => invoke('team_history_delete', { id })
+};
